create room

+++ laravel/database/migrations/2022_08_07_163718_create_rooms_table.php
@@ -15,7 +15,7 @@ class CreateRoomsTable extends Migration
     {
         Schema::create('rooms', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('room_id');
+            $table->bigIncrements('room_id');
             $table->integer('user_id');
             $table->timestamps();
         });
